requirements.txt
coolprop>=6.4.0
matplotlib>=3.5.0
pandas>=1.3.0
sqlalchemy>=1.4.0
pymysql>=1.0.0
numpy>=1.21.0
scipy>=1.7.0
python-dateutil>=2.8.0

